requirements.txt
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pandas>=2.1.0
python-dotenv>=1.0.0
gspread>=6.0.0
oauth2client>=4.1.3
requests>=2.31.0
